202202252010 - Commit
(Player.cpp) Added (S)pell (P)oints to the levelUp function. Which, is a copy of the strength increase for the same function

(Player.cpp) Added (S)pell (P)oints to the output stream in the battlePrint function

(Player.h/cpp)&&(Monster.h/cpp)->(Object.h/cpp) Moved isDead function to the more appropriate Object class as it is used by both the Player and Monster object. So, it would make more sense for the parent to have that ability

(Monster.cpp) Modified the Monster conversion constructor to implement the ACVariance for each individual Monster

(Item.h) Added default constructor to the Item class

(Object.cpp) Cleared a build error from the random seed and engine being marked as static in the header file and cpp file

Clean Build